pyproject.toml
[project]
name = "hubspot-mcp-server"
version = "0.1.0"
description = "MCP server for HubSpot CRM integration"
requires-python = ">=3.10"

dependencies = [
    "mcp",
    "requests",
    "pydantic",
    "python-dotenv>=1.1.1",
    "hubspot-api-client>=12.0.0",
    "httpx>=0.28.1",
    "fastmcp>=2.12.4",
    "aiohttp>=3.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
